parsing

+++ cpp04/ex03/Character.cpp
@@ -6,7 +6,7 @@
 /*   By: Leo Suardi <lsuardi@student.42.fr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/26 16:38:10 by Leo Suardi        #+#    #+#             */
-/*   Updated: 2022/01/23 15:56:15 by Leo Suardi       ###   ########.fr       */
+/*   Updated: 2022/01/23 18:02:36 by Leo Suardi       ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,14 +20,14 @@
 Character::Character() : m_name("unnamed"), m_itemCnt(0) { }
 
 Character::~Character() {
-	for (unsigned i = 0; i < m_itemCnt; ++i)
+	for (int i = 0; i < m_itemCnt; ++i)
 		delete m_inventory[i];
 }
 
 Character::Character(const Character &other)
 :	m_name(other.m_name),
 	m_itemCnt(other.m_itemCnt) {
-	for (unsigned i = 0; i < m_itemCnt; ++i)
+	for (int i = 0; i < m_itemCnt; ++i)
 		m_inventory[i] = other.m_inventory[i]->clone();
 }
 
@@ -36,7 +36,7 @@ Character::Character(std::string name) : m_name(name), m_itemCnt(0) { }
 Character	&Character::operator =(const Character &other) {
 	m_name = other.m_name;
 	m_itemCnt = other.m_itemCnt;
-	for (unsigned i = 0; i < m_itemCnt; ++i)
+	for (int i = 0; i < m_itemCnt; ++i)
 		m_inventory[i] = other.m_inventory[i]->clone();
 	return *this;
 }
@@ -49,13 +49,13 @@ void				Character::equip(AMateria *m) {
 }
 
 void				Character::unequip(int idx) {
-	if (idx < 0 || idx >= static_cast< int >(m_itemCnt))
+	if (idx < 0 || idx >= m_itemCnt)
 		throw std::out_of_range("Character::unequip");
 	ft::copy(m_inventory + idx + 1, m_inventory + m_itemCnt--, m_inventory + idx);
 }
 
 void				Character::use(int idx, ICharacter &target) {
-	if (idx < 0 || idx >= static_cast< int >(m_itemCnt))
+	if (idx < 0 || idx >= m_itemCnt)
 		throw std::out_of_range("Character::use");
 	m_inventory[idx]->use(target);
 }
